Bug 387557: Short circuit in the jar processor if we don't do any processing steps.
If we don't do any processing steps in the Jar Processor, just skip the
jar and leave the one we started with.

+++ bundles/org.eclipse.equinox.p2.jarprocessor/src/org/eclipse/internal/provisional/equinox/p2/jarprocessor/JarProcessor.java
@@ -339,6 +339,10 @@ public class JarProcessor {
 			//have to normalize after the post steps
 			normalize(result, workingDir);
 
+			// If the original input is where we ended up, just return it
+			if (input.equals(result))
+				return result;
+
 			if (!result.equals(workingFile) && !workingFile.equals(input))
 				workingFile.delete();
 			if (!result.getParentFile().equals(workingDir)) {
